refactor common.js
