+++ ui/config/config.js
@@ -10,7 +10,7 @@ import '../jobs/jobs_config.js';
 import '../oopsyraidsy/oopsyraidsy_config.js';
 import '../radar/radar_config.js';
 import '../raidboss/raidboss_config.js';
-import '../../resources/common.js';
+import '../../resources/common.ts';
 
 const Options = {};
 let gConfig = null;
