+++ ui/oopsyraidsy/oopsyraidsy.js
@@ -4,7 +4,7 @@ import NetRegexes from '../../resources/netregexes.js';
 import PartyTracker from '../../resources/party.js';
 import Regexes from '../../resources/regexes.js';
 import UserConfig from '../../resources/user_config.js';
-import { Util } from '../../resources/common.js';
+import * as Util from '../../resources/util';
 import ZoneId from '../../resources/zone_id.js';
 import ZoneInfo from '../../resources/zone_info.js';
