+++ ui/pullcounter/pullcounter.js
@@ -1,4 +1,4 @@
-import '../../resources/common.js';
+import '../../resources/common.ts';
 import { LocaleRegex } from '../../resources/translations.js';
 import Regexes from '../../resources/regexes.js';
 import UserConfig from '../../resources/user_config.js';
