+++ ui/raidboss/data/05-shb/raid/e10s.js
@@ -1,4 +1,4 @@
-import { Util } from '../../../../../resources/common.js';
+import { Util } from '../../../../../resources/common.ts';
 import Conditions from '../../../../../resources/conditions.js';
 import NetRegexes from '../../../../../resources/netregexes.js';
 import Outputs from '../../../../../resources/outputs.js';
