+++ ui/raidboss/emulator/data/network_log_converter/LineEvent0x03.js
@@ -1,6 +1,6 @@
 import LineEvent from './LineEvent.js';
 import EmulatorCommon from '../../EmulatorCommon.js';
-import { Util } from '../../../../../resources/common.js';
+import { Util } from '../../../../../resources/common.ts';
 
 // Added combatant event
 export class LineEvent0x03 extends LineEvent {
