+++ ui/raidboss/raidboss.js
@@ -1,5 +1,5 @@
 import './raidboss_config.js';
-import '../../resources/common.js';
+import '../../resources/common.ts';
 import '../../resources/timerbar.js';
 
 import { PopupText, PopupTextGenerator } from './popup-text.js';
